Vision
Stair coin plans to work with more partners to become part of the coin economy and become a major community coin in the future. When people buy or sell ladder money, there will be a decrease in total supply, the decrease in supply will occur with regular coin burning. This means that our circulating supply will continue to decline, while at the same time ladder money will become increasingly valuable. We have allocated a certain part of the supply to the "Green World" project. At the same time, we plan to make project payments as ladder money in the near future. Although our path is long and difficult, we aim to reach the top by working with perseverance and patience.